+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2022-01-26 - Edu
+### Added
+- Added **ambiguos redirections** error for empty expand redirections
+
 ## 2022-01-25 - Jorge
 ### Added
 - Files given with redirections are now checked to make sure they are suitable for each redirection.